+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - `lexicmap utils seed-pos`:
     - New columns: `sseqid` and `pos_seq`
     - More accurate seed distance.
+    - Add histograms of numbers of seed in sliding windows.
 - New commands:
     - **`lexicmap utils 2blast`: Convert the default search output to blast-style format**.